// src/app/models/member.ts
export class Member {
    Email:string='';
    FirstName: string = '';
    LastName: string = '';
    Female?: string = '';
    SkypeId?: string = '';
    ParentEmail?:string='';
    PhoneNumber: string = '';
    CreatedOn: Date;
    BirthDate: Date;
    Street?: string = '';
    ZipCode?: string = '';
    City?: string = '';
    Country?: string = '';
    Photo?: string = '';
    FullName: string='';
    Password: string;
    SchoolId: number;
    Login?:string;
    Status:number;
    State?:number
    UserName : string
    Id:string
}